Moved apparent_dip and azimuth_diff to analysis.py
