+++ mplstereonet/kinematic_analysis.py
@@ -14,57 +14,8 @@ import numpy as np
 from shapely import geometry, ops
 from . import stereonet_math 
 from .convenience_functions import subplots
-
-def apparent_dip(true_dip, azimuth_diff):
-    """
-    Calculate the apparent dip angle(s), given the true dip angle(s) and the 
-    azimuth (i.e. bearing) difference between the apparent dip direction(s) and
-    true dip direction(s). All in degrees.
-
-    Parameters
-    ----------
-    true_dip : number or sequence of numbers
-        true dip angle(s) in degrees
-    azimuth_diff : number or sequence of numbers
-        azimuth difference between the apparent dip direction(s) and true dip 
-        direction(s), in degrees
-
-    Returns
-    -------
-    apparent_dip : array
-        apparent dip angle(s) in degrees
-    """    
-    true_dip, azimuth_diff = np.atleast_1d(true_dip, azimuth_diff)
-    azimuth_diff = np.radians(np.abs(azimuth_diff))
-    true_dip = np.radians(true_dip)
-    apparent_dip = np.arctan(np.cos(azimuth_diff)*np.tan(true_dip))
-    return np.degrees(apparent_dip)
-
-def azimuth_diff(true_dip, apparent_dip):
-    """
-    Calculate the absolute difference in azimuth (i.e. bearing) between true 
-    dip direction(s) and apparent dip direction(s), given the true dip angle(s)
-    and the apparent dip angle(s). All in degrees.
-
-    Parameters
-    ----------
-    true_dip : number or sequence of numbers
-        true dip angle(s) in degrees
-    apparent_dip : number or sequence of numbers
-        apparent dip angle(s) in degrees
-
-    Returns
-    -------
-    azimuth_diff : array
-        azimuth difference between the apparent dip direction(s) and true dip 
-        direction(s)
-    """    
-    true_dip, apparent_dip = np.atleast_1d(true_dip, apparent_dip)
-    true_dip = np.radians(true_dip)
-    apparent_dip = np.radians(apparent_dip)
-    azimuth_diff = np.arccos(np.tan(apparent_dip)/np.tan(true_dip))
-    return np.degrees(azimuth_diff)
-    
+from .analysis import azimuth_diff, apparent_dip
+   
 def daylight_envelope(strike, dip, segments=500):
     """
     Calculates the longitude and latitude of `segments` points along the
